Add Tests for the Backend CRUD

+++ nonbonded/backend/database/crud/datasets.py
@@ -5,6 +5,7 @@ from nonbonded.backend.database.crud.authors import AuthorCRUD
 from nonbonded.backend.database.utilities.exceptions import (
     DataSetExistsError,
     DataSetNotFoundError,
+    UnableToDeleteError,
 )
 from nonbonded.library.models import datasets
 
@@ -119,5 +120,17 @@ class DataSetCRUD:
 
         if not db_data_set:
             raise DataSetNotFoundError(data_set_id)
+
+        if len(db_data_set.optimizations) > 0 or len(db_data_set.benchmarks) > 0:
+
+            type_name = (
+                "benchmark" if len(db_data_set.benchmarks) > 0 else "optimization"
+            )
+
+            raise UnableToDeleteError(
+                f"This data set (id={data_set_id}) is being referenced by at least "
+                f"one {type_name} and so cannot be deleted. Delete the referencing "
+                f"{type_name} first and then try again."
+            )
 
         db.delete(db_data_set)
